controller: rename controller
to peerpodconfig-ctrl
