+++ peerpodconfig-ctrl/Makefile
@@ -28,8 +28,8 @@ BUNDLE_METADATA_OPTS ?= $(BUNDLE_CHANNELS) $(BUNDLE_DEFAULT_CHANNEL)
 # This variable is used to construct full image tags for bundle and catalog images.
 #
 # For example, running 'make bundle-build bundle-push catalog-build catalog-push' will build and push both
-# confidentialcontainers.org/peer-pod-controller-bundle:$VERSION and confidentialcontainers.org/peer-pod-controller-catalog:$VERSION.
-IMAGE_TAG_BASE ?= confidentialcontainers.org/peer-pod-controller
+# confidentialcontainers.org/peerpodconfig-ctrl-bundle:$VERSION and confidentialcontainers.org/peerpodconfig-ctrl-catalog:$VERSION.
+IMAGE_TAG_BASE ?= confidentialcontainers.org/peerpodconfig-ctrl
 
 # BUNDLE_IMG defines the image:tag used for the bundle.
 # You can use it as an arg. (E.g make bundle-build BUNDLE_IMG=<some-registry>/<project-name-bundle>:<tag>)
