+++ peerpodconfig-ctrl/config/default/kustomization.yaml
@@ -1,12 +1,12 @@
 # Adds namespace to all resources.
-namespace: peer-pod-controller-system
+namespace: peerpodconfig-ctrl-system
 
 # Value of this field is prepended to the
 # names of all resources, e.g. a deployment named
 # "wordpress" becomes "alices-wordpress".
 # Note that it should also match with the prefix (text before '-') of the namespace
 # field above.
-namePrefix: peer-pod-controller-
+namePrefix: peerpodconfig-ctrl-
 
 # Labels to add to all resources and selectors.
 #commonLabels:
