+++ peerpodconfig-ctrl/config/manager/kustomization.yaml
@@ -12,5 +12,5 @@ apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 images:
 - name: controller
-  newName: quay.io/confidential-containers/peer-pod-controller
+  newName: quay.io/confidential-containers/peerpodconfig-ctrl
   newTag: 0.2.0
