+++ peerpodconfig-ctrl/config/manifests/kustomization.yaml
@@ -1,7 +1,7 @@
 # These resources constitute the fully configured set of manifests
 # used to generate the 'manifests/' directory in a bundle.
 resources:
-- bases/peer-pod-controller.clusterserviceversion.yaml
+- bases/peerpodconfig-ctrl.clusterserviceversion.yaml
 - ../default
 - ../samples
 - ../scorecard
